Tidy up the test code.
Replace SessionControllerFactory with the Builder written for the email
tests, and extract the `FromTemplate()` helper method to a separate
class so it can be shared across all email tests.
