+++ DDDEastAnglia.Tests/Admin/AdminHomeControllerTests.cs
@@ -1,5 +1,6 @@
 ﻿using DDDEastAnglia.Areas.Admin.Controllers;
 using DDDEastAnglia.Areas.Admin.Models;
+using DDDEastAnglia.Tests.Builders;
 using NUnit.Framework;
 
 namespace DDDEastAnglia.Tests.Admin
