+++ DDDEastAnglia.Tests/Admin/VotingControllerTests.cs
@@ -3,6 +3,7 @@ using DDDEastAnglia.Areas.Admin.Controllers;
 using DDDEastAnglia.Areas.Admin.Models;
 using DDDEastAnglia.DataAccess;
 using DDDEastAnglia.Helpers;
+using DDDEastAnglia.Tests.Builders;
 using DDDEastAnglia.VotingData;
 using DDDEastAnglia.VotingData.Models;
 using NSubstitute;
