+++ DDDEastAnglia.Tests/Controllers/HomeControllerTests.cs
@@ -1,5 +1,6 @@
 ﻿using DDDEastAnglia.Controllers;
 using DDDEastAnglia.Models;
+using DDDEastAnglia.Tests.Builders;
 using NUnit.Framework;
 
 namespace DDDEastAnglia.Tests.Controllers
