+++ DDDEastAnglia.Tests/Controllers/SessionControllerTests.cs
@@ -1,15 +1,11 @@
-﻿using System.Net.Mail;
-using DDDEastAnglia.Controllers;
-using DDDEastAnglia.DataAccess;
-using DDDEastAnglia.Domain;
-using DDDEastAnglia.Helpers;
-using DDDEastAnglia.Models;
+﻿using DDDEastAnglia.Models;
 using DDDEastAnglia.Services.Messenger.Email;
 using NSubstitute;
 using NUnit.Framework;
 using System.Web.Mvc;
 using DDDEastAnglia.Services.Messenger.Email.Templates;
-using MailMessage = DDDEastAnglia.Services.Messenger.Email.MailMessage;
+using DDDEastAnglia.Tests.Builders;
+using MailMessage = DDDEastAnglia.Tests.Helpers.Email.MailMessage;
 
 namespace DDDEastAnglia.Tests.Controllers
 {
@@ -19,9 +15,7 @@ namespace DDDEastAnglia.Tests.Controllers
         [Test]
         public void CannotSeeSessionList_WhenTheConferenceSaysThatSessionsCannotBeShown()
         {
-            var conference = Substitute.For<IConference>();
-            conference.CanShowSessions().Returns(false);
-            var controller = SessionControllerFactory.Create(conference);
+            var controller = new SessionControllerBuilder().Build();
 
             var result = controller.Index();
 
@@ -31,9 +25,7 @@ namespace DDDEastAnglia.Tests.Controllers
         [Test]
         public void CannotSubmitASession_WhenSessionSubmissionIsClosed()
         {
-            var conference = Substitute.For<IConference>();
-            conference.CanSubmit().Returns(false);
-            var controller = SessionControllerFactory.Create(conference);
+            var controller = new SessionControllerBuilder().Build();
 
             var result = controller.Create(new Session());
 
@@ -44,7 +36,7 @@ namespace DDDEastAnglia.Tests.Controllers
         public void CannotBeginToEditASession_WhenTheSessionBelongsToAnotherUser()
         {
             var session = new Session { SessionId = 1, SpeakerUserName = "bob" };
-            var controller = SessionControllerFactory.Create(session);
+            var controller = new SessionControllerBuilder().Updating(session).Build();
 
             var result = controller.Edit("fred", 1);
 
@@ -55,7 +47,7 @@ namespace DDDEastAnglia.Tests.Controllers
         public void CannotSaveAnEditToASession_WhenTheSessionBelongsToAnotherUser()
         {
             var session = new Session { SessionId = 1, SpeakerUserName = "bob" };
-            var controller = SessionControllerFactory.Create(session);
+            var controller = new SessionControllerBuilder().Updating(session).Build();
 
             var result = controller.Edit("fred", session);
 
@@ -66,7 +58,7 @@ namespace DDDEastAnglia.Tests.Controllers
         public void CannotBeginToDeleteASession_WhenTheSessionBelongsToAnotherUser()
         {
             var session = new Session { SessionId = 1, SpeakerUserName = "bob" };
-            var controller = SessionControllerFactory.Create(session);
+            var controller = new SessionControllerBuilder().Updating(session).Build();
 
             var result = controller.Delete("fred", 1);
 
@@ -77,7 +69,7 @@ namespace DDDEastAnglia.Tests.Controllers
         public void CannotConfirmToDeleteASession_WhenTheSessionBelongsToAnotherUser()
         {
             var session = new Session { SessionId = 1, SpeakerUserName = "bob" };
-            var controller = SessionControllerFactory.Create(session);
+            var controller = new SessionControllerBuilder().Updating(session).Build();
 
             var result = controller.DeleteConfirmed("fred", 1);
 
@@ -108,7 +100,7 @@ namespace DDDEastAnglia.Tests.Controllers
 
                 controller.Create(session);
 
-                var expectedMailMessage = FromTemplate(SessionCreatedMailTemplate.Create(session), bob);
+                var expectedMailMessage = MailMessage.FromTemplate(SessionCreatedMailTemplate.Create(session), bob);
                 postman.Received().Deliver(expectedMailMessage);
             }
 
@@ -132,106 +124,8 @@ namespace DDDEastAnglia.Tests.Controllers
 
                 controller.Edit(bob.UserName, session);
 
-                var expectedMailMessage = FromTemplate(SessionUpdatedMailTemplate.Create(session), bob);
+                var expectedMailMessage = MailMessage.FromTemplate(SessionUpdatedMailTemplate.Create(session), bob);
                 postman.Received().Deliver(expectedMailMessage);
-            }
-
-            private static MailMessage FromTemplate(IMailTemplate mailTemplate, UserProfile userProfile)
-            {
-                return new MailMessage
-                {
-                    To = new MailAddress(userProfile.EmailAddress, userProfile.Name),
-                    From = new MailAddress("admin@dddeastanglia.com", "DDD East Anglia"),
-                    Subject = mailTemplate.RenderSubjectLine(),
-                    Body = mailTemplate.RenderBody()
-                };
-            }
-        }
-
-        private class SessionControllerBuilder
-        {
-            private IConferenceLoader conferenceLoader;
-            private IUserProfileRepository userProfileRepository;
-            private ISessionRepository sessionRepository;
-            private ISessionSorter sessionSorter;
-            private IPostman postman;
-            private UserProfile user;
-
-            public SessionControllerBuilder()
-            {
-                conferenceLoader = Substitute.For<IConferenceLoader>();
-                userProfileRepository = Substitute.For<IUserProfileRepository>();
-                sessionRepository = Substitute.For<ISessionRepository>();
-                sessionSorter = Substitute.For<ISessionSorter>();
-                postman = Substitute.For<IPostman>();
-            }
-
-            public SessionControllerBuilder WithPostman(IPostman newPostman)
-            {
-                postman = newPostman;
-                return this;
-            }
-
-            public SessionControllerBuilder WhenSubmissionsAreOpen()
-            {
-                var conference = Substitute.For<IConference>();
-                conference.CanSubmit().Returns(true);
-                conferenceLoader.LoadConference().Returns(conference);
-
-                return this;
-            }
-
-            public SessionControllerBuilder ForUser(UserProfile newUser)
-            {
-                user = newUser;
-                userProfileRepository.GetUserProfileByUserName(newUser.UserName).Returns(newUser);
-
-                return this;
-            }
-
-            public SessionControllerBuilder Submitting(Session session)
-            {
-                sessionRepository.AddSession(session).Returns(session);
-                return this;
-            }
-
-            public SessionControllerBuilder Updating(Session session)
-            {
-                sessionRepository.Get(session.SessionId).Returns(session);
-                return this;
-            }
-
-            public SessionController Build()
-            {
-                var sessionController = new SessionController(
-                    conferenceLoader,
-                    userProfileRepository,
-                    sessionRepository,
-                    sessionSorter,
-                    new EmailMessengerFactory(postman));
-
-                sessionController.SetupWithAuthenticatedUser(user);
-
-                return sessionController;
-            }
-        }
-
-        private static class SessionControllerFactory
-        {
-            public static SessionController Create(IConference conference)
-            {
-                var conferenceLoader = Substitute.For<IConferenceLoader>();
-                conferenceLoader.LoadConference().Returns(conference);
-                var controller = new SessionController(conferenceLoader, Substitute.For<IUserProfileRepository>(), Substitute.For<ISessionRepository>(), Substitute.For<ISessionSorter>(), new EmailMessengerFactory(Substitute.For<IPostman>()));
-                return controller;
-            }
-
-            public static SessionController Create(Session session)
-            {
-                var sessionRepository = Substitute.For<ISessionRepository>();
-                sessionRepository.Get(session.SessionId).Returns(session);
-                var controller = new SessionController(Substitute.For<IConferenceLoader>(), Substitute.For<IUserProfileRepository>(), sessionRepository, Substitute.For<ISessionSorter>(), new EmailMessengerFactory(Substitute.For<IPostman>()));
-                return controller;
             }
         }
     }
