+++ DDDEastAnglia.Tests/DDDEastAnglia.Tests.csproj
@@ -60,20 +60,22 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Admin\AdminHomeControllerTests.cs" />
-    <Compile Include="ChartDataConverterBuilder.cs" />
-    <Compile Include="DataProviderBuilder.cs" />
+    <Compile Include="Builders\ChartDataConverterBuilder.cs" />
+    <Compile Include="Builders\SessionControllerBuilder.cs" />
+    <Compile Include="Builders\DataProviderBuilder.cs" />
     <Compile Include="Admin\SessionControllerTests.cs" />
     <Compile Include="Admin\UserControllerTests.cs" />
-    <Compile Include="DnsLookupBuilder.cs" />
+    <Compile Include="Builders\DnsLookupBuilder.cs" />
+    <Compile Include="Helpers\Email\MailMessage.cs" />
     <Compile Include="Helpers\Email\SendGrid\HtmlRendererFacts.cs" />
     <Compile Include="Services\Messengers\Email\SessionSubmissionMessengerFacts.cs" />
     <Compile Include="Services\Messengers\Email\Templates\SessionUpdatedMailTemplateFacts.cs" />
     <Compile Include="Services\Messengers\Email\Templates\SessionCreatedMailTemplateFacts.cs" />
-    <Compile Include="VotingControllerBuilder.cs" />
+    <Compile Include="Builders\VotingControllerBuilder.cs" />
     <Compile Include="Admin\VotingControllerTests.cs" />
     <Compile Include="CannedResponseDateTimeOffsetProvider.cs" />
     <Compile Include="ConferenceHelper.cs" />
-    <Compile Include="ConferenceLoaderBuilder.cs" />
+    <Compile Include="Builders\ConferenceLoaderBuilder.cs" />
     <Compile Include="ActionResultExtensions.cs" />
     <Compile Include="Controllers\TimelineControllerTests.cs" />
     <Compile Include="Controllers\HomeControllerTests.cs" />
